gsoc-i18n: svn merge -r 15605:15894 http://svn.civicrm.org/civicrm/trunk
git-svn-id: http://svn.civicrm.org/civicrm/branches/gsoc-i18n@15895 cff41f8d-9fe8-0310-9d4e-fb156a3ffe5e

+++ CRM/Core/BAO/File.php
@@ -94,7 +94,8 @@ class CRM_Core_BAO_File extends CRM_Core_DAO_File {
                                     $entityTable, $entityID,
                                     $entitySubtype, $overwrite = true,
                                     $fileParams = null,
-                                    $uploadName = 'uploadFile' ) {
+                                    $uploadName = 'uploadFile',
+                                    $mimeType   = null ) {
 
         require_once 'CRM/Core/DAO/File.php';
         $config = & CRM_Core_Config::singleton();
@@ -126,12 +127,15 @@ class CRM_Core_BAO_File extends CRM_Core_DAO_File {
 
         $dao =& CRM_Core_DAO::executeQuery( $sql, $params );
         $dao->fetch();
-       
-        $mimeType = $_FILES[$uploadName]['type'];
+
+        if ( ! $mimeType ) {
+            $mimeType = $_FILES[$uploadName]['type'];
+        }
         
         require_once "CRM/Core/DAO/File.php";
         $fileDAO =& new CRM_Core_DAO_File();
-        if ( $dao->cfID ) {
+        if ( isset( $dao->cfID ) &&
+             $dao->cfID ) {
             $fileDAO->id = $dao->cfID;
             unlink( $directoryName . DIRECTORY_SEPARATOR . $dao->uri );
         }
@@ -149,7 +153,8 @@ class CRM_Core_BAO_File extends CRM_Core_DAO_File {
         // need to add/update civicrm_entity_file
         require_once "CRM/Core/DAO/EntityFile.php";
         $entityFileDAO =& new CRM_Core_DAO_EntityFile();
-        if ($dao->cefID ) {
+        if ( isset( $dao->cefID ) &&
+             $dao->cefID ) {
             $entityFileDAO->id =  $dao->cefID;
         }
         $entityFileDAO->entity_table = $entityTable;
@@ -226,20 +231,21 @@ class CRM_Core_BAO_File extends CRM_Core_DAO_File {
      * combination
      */
     public function &getEntityFile( $entityTable, $entityID ) {
+        require_once 'CRM/Utils/File.php';
         $config =& CRM_Core_Config::singleton( );
 
         list( $sql, $params ) = self::sql( $entityTable, $entityID, null );
         $dao    = CRM_Core_DAO::executeQuery( $sql, $params );
         $results = array( );
         while ( $dao->fetch( ) ) {
-            $result['fileID'  ]  = $dao->cfID;
-            $result['entityID']  = $dao->cefID;
+            $result['fileID'   ] = $dao->cfID;
+            $result['entityID' ] = $dao->cefID;
             $result['mime_type'] = $dao->mime_type;
-            $result['fileName']  = $dao->uri;
-            $result['cleanName'] = self::cleanName( $dao->uri );
-            $result['fullPath']  = $config->customFileUploadDir . DIRECTORY_SEPARATOR . $dao->uri;
-            $result['url'     ]  = CRM_Utils_System::url( 'civicrm/file', "reset=1&id={$dao->cfID}&eid={$entityID}" );
-            $result['href'    ]  = "<a href=\"{$result['url']}\">{$dao->uri}</a>";
+            $result['fileName' ] = $dao->uri;
+            $result['cleanName'] = CRM_Utils_File::cleanFileName( $dao->uri );
+            $result['fullPath' ] = $config->customFileUploadDir . DIRECTORY_SEPARATOR . $dao->uri;
+            $result['url'      ] = CRM_Utils_System::url( 'civicrm/file', "reset=1&id={$dao->cfID}&eid={$entityID}" );
+            $result['href'     ] = "<a href=\"{$result['url']}\">{$result['cleanName']}</a>";
             $results[$dao->cfID] = $result;
         }
         return $results;
@@ -267,15 +273,103 @@ AND       CEF.entity_id    = %2";
 
     }
     
-    /**
-     * remove the 32 bit md5 we add to the fileName
-     * also remove the unknown tag if we added it
-     */
-    function cleanName( $name ) {
-        // replace the last 33 character before the '.' with null
-        $name = preg_replace( '/(_[\w]{32})\./', '.', $name );
-        return $name;
+    static function buildAttachment( &$form, $entityTable, $entityID = null, $numAttachments = 3 ) {
+
+        // add 3 attachments
+        for ( $i = 1; $i <= $numAttachments; $i++ ) {
+            $form->addElement( 'file', "attachFile_$i", ts('Attach File'), 'size=30 maxlength=60' );
+            $form->setMaxFileSize( 2 * 1024 * 1024 );
+            $form->addRule( "attachFile_$i",
+                            ts( 'File size should be less than %1 MByte(s)',
+                                array( 1 => 2 ) ),
+                            'maxfilesize',
+                            2 * 1024 * 1024 );
+        }
+
+        $attachmentInfo = self::attachmentInfo(  $entityTable, $entityID );
+        if ( $attachmentInfo ) {
+            $form->add( 'checkbox', 'is_delete_attachment', ts( 'Delete Current Attachment(s)' ) );
+            $form->assign( 'currentAttachmentURL',
+                           $attachmentInfo );
+        } else {
+            $form->assign( 'currentAttachmentURL', null );
+        }
     }
+
+    static function attachmentInfo( $entityTable, $entityID, $separator = '<br />' ) {
+        if ( ! $entityID ) {
+            return null;
+        }
+
+        $currentAttachments = self::getEntityFile( $entityTable,
+                                                   $entityID );
+        if ( ! empty( $currentAttachments ) ) {
+            $currentAttachmentURL = array( );
+            foreach ( $currentAttachments as $fileID => $attach ) {
+                $currentAttachmentURL[] = $attach['href'];
+            }
+            return implode( $separator, $currentAttachmentURL );
+        }
+        return null;
+    }
+
+    static function formatAttachment( &$formValues,
+                                      &$params,
+                                      $entityTable,
+                                      $entityID = null,
+                                      $numAttachments = 3 ) {
+
+        // delete current attachments if applicable
+        if ( $entityID &&
+             CRM_Utils_Array::value( 'is_delete_attachment', $formValues ) ) {
+            CRM_Core_BAO_File::deleteEntityFile( $entityTable,
+                                                 $entityID );
+        }
+
+        // setup all attachments
+        for ( $i = 1; $i <= $numAttachments; $i++ ) {
+            $attachName = "attachFile_$i";
+            if ( isset( $formValues[$attachName] ) &&
+                 ! empty( $formValues[$attachName] ) ) {
+                // ensure file is not empty
+                $contents = file_get_contents( $formValues[$attachName] );
+                if ( $contents ) {
+                    $fileParams = array( 'uri'        => $_FILES[$attachName]['name'],
+                                         'type'       => $_FILES[$attachName]['type'],
+                                         'upload_date'=> date( 'Ymdhis' ),
+                                         'location'   => $formValues[$attachName] );
+                    $params[$attachName] = $fileParams;
+                }
+            }
+        }
+    }
+
+    static function processAttachment( &$params,
+                                       $entityTable,
+                                       $entityID,
+                                       $numAttachments = 3 ) {
+        for ( $i = 1; $i <= $numAttachments; $i++ ) {
+            if ( isset( $params["attachFile_$i"] ) &&
+                 is_array( $params["attachFile_$i"] ) ) {
+                self::filePostProcess($params["attachFile_$i"]['location'],
+                                      null, 
+                                      $entityTable,
+                                      $entityID,
+                                      null,
+                                      true,
+                                      $params["attachFile_$i"],
+                                      "attachFile_$i",
+                                      $params["attachFile_$i"]['type'] );
+            }
+        }
+    }
+
+    static function uploadNames( ) {
+        return array( 'attachFile_1',
+                      'attachFile_2',
+                      'attachFile_3' );
+    }
+
 
 }
